requirements.txt
discord.py
python-telegram-bot[ext]
python-dotenv
